refactor: move test util out of `src`
